+++ tests/local/sris.php
@@ -8,5 +8,8 @@ require_once dirname(__FILE__, 2).'/includes/local.php';
 
 /* ------------------------------------------------------------------------------------------------------------------ */
 
-echo 'sha256-KbfTjB0WZ8vvXngdpJGY3Yp3xKk+tttbqClO11anCIU='."\n";
-echo c::sri('https://cdnjs.cloudflare.com/ajax/libs/highlight.js/9.9.0/highlight.min.js');
+echo 'sha256-jHqKZLWQrHz5TSDtGB1wqHlUcYHjK4XUN3IS84M9Av0='."\n";
+
+c::benchStart();
+echo c::sri('https://raw.githubusercontent.com/websharks/ace-builds/fork-1.2.6/src-min-noconflict/ext-spellcheck.js');
+c::benchPrint();